Update/1.95.1 (#61)
* Bump version

* Check if user exists before creating

* Better health checks

* Cleaning

* Update readme

* manifest update

* fix typo

+++ docker_entrypoint.sh
@@ -100,10 +100,14 @@ else
     echo "Synapse-admin user not found. Creating ..."
     echo
     admin_password=$(cat /dev/urandom | base64 | head -c 16)
-    timeout 210s /start.py &
-    sleep 205
-    register_new_matrix_user --config /data/homeserver.yaml --user admin --password $admin_password --admin
-    echo $admin_password > /data/start9/adm.key
+    timeout 25s /start.py &
+    sleep 20
+    admin_exists=$(sqlite3 /data/homeserver.db "SELECT name FROM users WHERE name LIKE '@admin%';")
+    if [ -z "$admin_exists" ]; then
+        # Create the admin user and key
+        register_new_matrix_user --config /data/homeserver.yaml --user admin --password $admin_password --admin
+        echo $admin_password > /data/start9/adm.key
+    fi
     python /configurator.py
 fi
